Make IPA e2e work with 32 bit primes
