+++ src/query/executor.rs
@@ -1,24 +1,18 @@
 use crate::{
-    error::Error,
-    ff::{Field, FieldType, Fp31, GaloisField, PrimeField, Serializable},
+    ff::{Field, GaloisField, Serializable},
     helpers::{
         negotiate_prss,
-        query::{IpaQueryConfig, QueryConfig, QueryType},
-        AlignedByteArrStream, ByteArrStream, Gateway, TotalRecords,
+        query::{QueryConfig, QueryType},
+        ByteArrStream, Gateway,
     },
     protocol::{
-        attribution::input::MCAggregateCreditOutputRow,
-        context::SemiHonestContext,
-        ipa::{ipa, IPAInputRow},
-        step::StepNarrow,
-        BreakdownKey, GenericStep, MatchKey,
+        attribution::input::MCAggregateCreditOutputRow, context::SemiHonestContext,
+        step::StepNarrow, GenericStep,
     },
-    secret_sharing::{
-        replicated::semi_honest::AdditiveShare as Replicated, Linear as LinearSecretSharing,
-    },
+    query::runner::IpaRunner,
+    secret_sharing::{replicated::semi_honest::AdditiveShare, Linear as LinearSecretSharing},
     task::JoinHandle,
 };
-use futures_util::StreamExt;
 use generic_array::GenericArray;
 use rand::rngs::StdRng;
 use rand_core::SeedableRng;
@@ -31,16 +25,16 @@ pub trait Result: Send + Debug {
     fn into_bytes(self: Box<Self>) -> Vec<u8>;
 }
 
-impl<F: Field> Result for Vec<Replicated<F>>
+impl<F: Field> Result for Vec<AdditiveShare<F>>
 where
-    Replicated<F>: Serializable,
+    AdditiveShare<F>: Serializable,
 {
     fn into_bytes(self: Box<Self>) -> Vec<u8> {
-        let mut r = vec![0u8; self.len() * <Replicated<F> as Serializable>::Size::USIZE];
+        let mut r = vec![0u8; self.len() * <AdditiveShare<F> as Serializable>::Size::USIZE];
         for (i, share) in self.into_iter().enumerate() {
             share.serialize(GenericArray::from_mut_slice(
-                &mut r[i * <Replicated<F> as Serializable>::Size::USIZE
-                    ..(i + 1) * <Replicated<F> as Serializable>::Size::USIZE],
+                &mut r[i * <AdditiveShare<F> as Serializable>::Size::USIZE
+                    ..(i + 1) * <AdditiveShare<F> as Serializable>::Size::USIZE],
             ));
         }
 
@@ -66,59 +60,6 @@ where
     }
 }
 
-#[cfg(any(test, feature = "cli", feature = "test-fixture"))]
-async fn execute_test_multiply<F: Field>(
-    ctx: SemiHonestContext<'_>,
-    mut input: AlignedByteArrStream,
-) -> std::result::Result<Vec<Replicated<F>>, Error>
-where
-    Replicated<F>: Serializable,
-{
-    use crate::protocol::{basics::SecureMul, RecordId};
-
-    let mut results = Vec::new();
-    while let Some(v) = input.next().await {
-        // multiply pairs
-        let mut a = None;
-        let mut record_id = 0_u32;
-        for share in Replicated::<F>::from_byte_slice(&v.unwrap()) {
-            match a {
-                None => a = Some(share),
-                Some(a_v) => {
-                    let result = a_v
-                        .multiply(&share, ctx.clone(), RecordId::from(record_id))
-                        .await
-                        .unwrap();
-                    results.push(result);
-                    record_id += 1;
-                    a = None;
-                }
-            }
-        }
-
-        assert!(a.is_none());
-    }
-
-    Ok(results)
-}
-
-async fn execute_ipa<F: PrimeField, MK: GaloisField, BK: GaloisField>(
-    ctx: SemiHonestContext<'_>,
-    query_config: IpaQueryConfig,
-    mut input: AlignedByteArrStream,
-) -> std::result::Result<Vec<MCAggregateCreditOutputRow<F, Replicated<F>, BK>>, Error>
-where
-    IPAInputRow<F, MK, BK>: Serializable,
-    Replicated<F>: Serializable,
-{
-    let mut input_vec = Vec::new();
-    while let Some(data) = input.next().await {
-        input_vec.extend(IPAInputRow::<F, MK, BK>::from_byte_slice(&data.unwrap()));
-    }
-
-    ipa(ctx, input_vec.as_slice(), query_config).await
-}
-
 #[allow(unused)]
 pub fn start_query(
     config: QueryConfig,
@@ -131,39 +72,19 @@ pub fn start_query(
         // Negotiate PRSS first
         let step = GenericStep::default().narrow(&config.query_type);
         let prss = negotiate_prss(&gateway, &step, &mut rng).await.unwrap();
+        let ctx = SemiHonestContext::new(&prss, &gateway);
 
-        match config.field_type {
-            FieldType::Fp31 => match config.query_type {
-                #[cfg(any(test, feature = "cli", feature = "test-fixture"))]
-                QueryType::TestMultiply => {
-                    let ctx = SemiHonestContext::new_with_total_records(
-                        &prss,
-                        &gateway,
-                        TotalRecords::Indeterminate,
-                    );
-                    let input = input.align(<Replicated<Fp31> as Serializable>::Size::USIZE);
-                    Box::new(execute_test_multiply::<Fp31>(ctx, input).await.unwrap())
-                        as Box<dyn Result>
-                }
-                QueryType::Ipa(config) => {
-                    let ctx = SemiHonestContext::new_with_total_records(
-                        &prss,
-                        &gateway,
-                        // will be specified in downstream steps
-                        TotalRecords::Unspecified,
-                    );
-                    let input = input.align(
-                        <IPAInputRow<Fp31, MatchKey, BreakdownKey> as Serializable>::Size::USIZE,
-                    );
-                    Box::new(
-                        execute_ipa::<Fp31, MatchKey, BreakdownKey>(ctx, config, input)
-                            .await
-                            .unwrap(),
-                    ) as Box<dyn Result>
-                }
-            },
-            FieldType::Fp32BitPrime => {
-                todo!()
+        match config.query_type {
+            #[cfg(any(test, feature = "cli", feature = "test-fixture"))]
+            QueryType::TestMultiply => {
+                super::runner::TestMultiplyRunner
+                    .run(ctx, config.field_type, input)
+                    .await
+            }
+            QueryType::Ipa(ipa_query_config) => {
+                IpaRunner(ipa_query_config)
+                    .run(ctx, config.field_type, input)
+                    .await
             }
         }
     })
@@ -172,117 +93,7 @@ pub fn start_query(
 #[cfg(all(test, not(feature = "shuttle"), feature = "in-memory-infra"))]
 mod tests {
     use super::*;
-    use crate::{
-        ipa_test_input,
-        protocol::context::Context,
-        secret_sharing::IntoShares,
-        test_fixture::{input::GenericReportTestInput, join3v, Reconstruct, TestWorld},
-    };
-    use generic_array::GenericArray;
-    use typenum::Unsigned;
-
-    #[tokio::test]
-    async fn multiply() {
-        let world = TestWorld::default();
-        let contexts = world
-            .contexts()
-            .map(|ctx| ctx.set_total_records(TotalRecords::Indeterminate));
-        let a = [Fp31::truncate_from(4u128), Fp31::truncate_from(5u128)];
-        let b = [Fp31::truncate_from(3u128), Fp31::truncate_from(6u128)];
-
-        let helper_shares = (a, b).share().map(|(a, b)| {
-            const SIZE: usize = <Replicated<Fp31> as Serializable>::Size::USIZE;
-            let r = a
-                .into_iter()
-                .zip(b)
-                .flat_map(|(a, b)| {
-                    let mut slice = [0_u8; 2 * SIZE];
-                    a.serialize(GenericArray::from_mut_slice(&mut slice[..SIZE]));
-                    b.serialize(GenericArray::from_mut_slice(&mut slice[SIZE..]));
-
-                    slice
-                })
-                .collect::<Vec<_>>();
-
-            ByteArrStream::from(r).align(SIZE)
-        });
-
-        let results = join3v(
-            helper_shares
-                .into_iter()
-                .zip(contexts)
-                .map(|(shares, context)| execute_test_multiply::<Fp31>(context, shares)),
-        )
-        .await;
-
-        let results = results.reconstruct();
-
-        assert_eq!(
-            vec![Fp31::truncate_from(12u128), Fp31::truncate_from(30u128)],
-            results
-        );
-    }
-
-    #[tokio::test]
-    async fn ipa() {
-        const EXPECTED: &[[u128; 2]] = &[[0, 0], [1, 2], [2, 3]];
-
-        let records: Vec<GenericReportTestInput<Fp31, MatchKey, BreakdownKey>> = ipa_test_input!(
-            [
-                { timestamp: 0, match_key: 12345, is_trigger_report: 0, breakdown_key: 1, trigger_value: 0 },
-                { timestamp: 0, match_key: 12345, is_trigger_report: 0, breakdown_key: 2, trigger_value: 0 },
-                { timestamp: 0, match_key: 68362, is_trigger_report: 0, breakdown_key: 1, trigger_value: 0 },
-                { timestamp: 0, match_key: 12345, is_trigger_report: 1, breakdown_key: 0, trigger_value: 5 },
-                { timestamp: 0, match_key: 68362, is_trigger_report: 1, breakdown_key: 0, trigger_value: 2 },
-            ];
-            (Fp31, MatchKey, BreakdownKey)
-        );
-        let records = records
-            .share()
-            // TODO: a trait would be useful here to convert IntoShares<T> to IntoShares<Vec<u8>>
-            .map(|shares| {
-                shares
-                    .into_iter()
-                    .flat_map(|share: IPAInputRow<Fp31, MatchKey, BreakdownKey>| {
-                        let mut buf = [0u8; <IPAInputRow<
-                            Fp31,
-                            MatchKey,
-                            BreakdownKey,
-                        > as Serializable>::Size::USIZE];
-                        share.serialize(GenericArray::from_mut_slice(&mut buf));
-
-                        buf
-                    })
-                    .collect::<Vec<_>>()
-            });
-
-        let world = TestWorld::default();
-        let contexts = world.contexts();
-        let results = join3v(records.into_iter().zip(contexts).map(|(shares, ctx)| {
-            let query_config = IpaQueryConfig {
-                num_multi_bits: 3,
-                per_user_credit_cap: 3,
-                attribution_window_seconds: 0,
-                max_breakdown_key: 3,
-            };
-            let input = ByteArrStream::from(shares)
-                .align(<IPAInputRow<Fp31, MatchKey, BreakdownKey> as Serializable>::Size::USIZE);
-            execute_ipa::<Fp31, MatchKey, BreakdownKey>(ctx, query_config, input)
-        }))
-        .await;
-
-        let results: Vec<GenericReportTestInput<Fp31, MatchKey, BreakdownKey>> =
-            results.reconstruct();
-        for (i, expected) in EXPECTED.iter().enumerate() {
-            assert_eq!(
-                *expected,
-                [
-                    results[i].breakdown_key.as_u128(),
-                    results[i].trigger_value.as_u128()
-                ]
-            );
-        }
-    }
+    use crate::{ff::Fp31, secret_sharing::IntoShares};
 
     #[test]
     fn serialize_result() {
@@ -294,7 +105,7 @@ mod tests {
         let bytes = Box::new(input).into_bytes();
         assert_eq!(
             expected,
-            Replicated::<Fp31>::from_byte_slice(&bytes).collect::<Vec<_>>()
+            AdditiveShare::<Fp31>::from_byte_slice(&bytes).collect::<Vec<_>>()
         );
     }
 }
