+++ src/query/mod.rs
@@ -1,5 +1,6 @@
 mod executor;
 mod processor;
+mod runner;
 mod state;
 
 pub use executor::Result as ProtocolResult;
